requirements.txt
pandas
mysql-connector-python
matplotlib
seaborn
scikit-learn
prophet
